project/requirements.txt
alabaster==0.7.13
asgiref==3.7.2
Babel==2.13.1
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
docutils==0.20.1
idna==3.6
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
psycopg2==2.9.9
Pygments==2.17.2
pytz==2023.3.post1
requests==2.31.0
snowballstemmer==2.2.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0